Move DataSet to its own file
Avoiding circular import for DataArray -> DataView -> DataSet (in DA)
